+++ R/vctrs-deprecated.R
@@ -49,3 +49,29 @@ vec_type2 <- function(x, y, ...) {
   warn_deprecated(c("`vec_type2()` has been renamed to `vec_ptype2()`."))
   vec_ptype2(x, y, ...)
 }
+
+#' Convert to an index vector
+#'
+#' @description
+#'
+#' \Sexpr[results=rd, stage=render]{vctrs:::lifecycle("soft-deprecated")}
+#'
+#' `vec_as_index()` has been renamed to [vec_as_location()] and is
+#' soft-deprecated as of vctrs 0.2.2.
+#'
+#' @inheritParams vec_as_location
+#'
+#' @keywords internal
+#' @export
+vec_as_index <- function(i, n, names = NULL) {
+  signal_soft_deprecated(paste_line(
+    "`vec_as_index()` is deprecated as of vctrs 0.2.2.",
+    "Please use `vec_as_location() instead.`"
+  ))
+  n <- vec_coercible_cast(n, integer())
+  vec_assert(n, integer(), 1L)
+  i <- vec_as_subscript(i)
+
+  convert_values <- as_opts_location_convert_values("negative")
+  .Call(vctrs_as_location, i, n, names, convert_values)
+}